Move shared ShimmerBox to ui/compose/components
Relocate the shared ShimmerBox composable from postsrs/screens to the shared
compose components package, alongside the other reusable UI primitives.
